Fix build on older compilers that do not support 'd' suffix.
Things like:  if (foo > 0.0d) { }
fails to compile on Fedora 13 (at least) because the compiler
does not support a 'd' suffix properly.

Use a slightly uglier explicit cast to double to work around
this compiler limitation.

+++ hostapd/config_file.c
@@ -2945,8 +2945,8 @@ static int hostapd_config_fill(struct hostapd_config *conf,
 			char *end;					\
 									\
 			conf->_val = strtod(pos, &end);			\
-			if (*end || conf->_val < 0.0d ||		\
-			    conf->_val > 1.0d) {			\
+			if (*end || conf->_val < (double)(0.0) ||	\
+			    conf->_val > (double)(1.0)) {		\
 				wpa_printf(MSG_ERROR,			\
 					   "Line %d: Invalid value '%s'", \
 					   line, pos);			\
